requirements.txt
anyio==4.12.0
AnyQt==0.2.1
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
baycomp==1.0.3
Bottleneck==1.6.0
build==1.3.0
cattrs==25.3.0
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
commonmark==0.9.1
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
dictdiffer==0.9.0
docutils==0.22.4
et_xmlfile==2.0.0
executing==2.2.1
fonttools==4.61.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jaraco.classes==3.4.0
jaraco.context==6.0.2
jaraco.functools==4.4.0
jedi==0.19.2
joblib==1.5.3
jupyter_client==8.7.0
jupyter_core==5.9.1
keyring==25.7.0
keyrings.alt==5.0.2
kiwisolver==1.4.9
matplotlib==3.10.8
matplotlib-inline==0.2.1
more-itertools==10.8.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.3.5
openpyxl==3.1.5
openTSNE==1.0.4
orange-canvas-core==0.2.7
orange-widget-base==4.27.0
Orange3==3.40.0
-e git+ssh://git@github.com/likit/OrangeWidgetsLabSimulator.git@b4d09f87b464b9d8d22e32fe5ab2c7cdf5b0d973#egg=Orange3_LabDES
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.1
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.2
pyqtgraph==0.14.0
python-dateutil==2.9.0.post0
python-louvain==0.16
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qasync==0.28.0
qtconsole==5.7.0
QtPy==2.4.3
requests==2.32.5
requests-cache==1.2.1
scikit-learn==1.8.0
scipy==1.16.3
serverfiles==0.3.1
setuptools==80.9.0
simpy==4.1.1
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.4
traitlets==5.14.3
truststore==0.10.4
typing_extensions==4.15.0
tzdata==2025.3
url-normalize==2.2.1
urllib3==2.6.2
wcwidth==0.2.14
wheel==0.45.1
xgboost==2.0.3
xlrd==2.0.2
xlsxwriter==3.2.9
